requirements.txt
Babel==2.10.1
google_api_python_client==2.50.0
mysql_connector_repackaged==0.3.1
protobuf==4.21.1
python-dotenv==0.20.0
python_dateutil==2.8.2
python_telegram_bot==13.11
requests==2.25.1
schedule==1.1.0
telegram==0.0.1
